pyproject.toml
[tool.poetry]
name = "turma2"
version = "0.1.0"
description = ""
authors = ["Guilherme Pupio <guicpupio@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = {extras = ["job-queue"], version = "^20.7"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
